fix according to comments

+++ state/carmen_test.go
@@ -132,7 +132,7 @@ func TestCarmenState_AccountBalanceOperations(t *testing.T) {
 
 			// get randomized balance
 			additionBase := GetRandom(1, 1000*5000)
-			addition := uint256.NewInt(uint64(additionBase))
+			addition := uint256.NewInt(additionBase)
 
 			csDB.AddBalance(addr, addition, 0)
 
@@ -140,7 +140,7 @@ func TestCarmenState_AccountBalanceOperations(t *testing.T) {
 				t.Fatal("failed to add balance to carmen state DB account")
 			}
 
-			subtraction := uint256.NewInt(uint64(GetRandom(1, additionBase)))
+			subtraction := uint256.NewInt(GetRandom(1, int(additionBase)))
 			expectedResult := uint256.NewInt(0).Sub(addition, subtraction)
 
 			csDB.SubBalance(addr, subtraction, 0)
@@ -178,7 +178,7 @@ func TestCarmenState_NonceOperations(t *testing.T) {
 			csDB.CreateAccount(addr)
 
 			// get randomized nonce
-			newNonce := uint64(GetRandom(1, 1000*5000))
+			newNonce := GetRandom(1, 1000*5000)
 
 			csDB.SetNonce(addr, newNonce)
 
@@ -349,7 +349,7 @@ func TestCarmenState_RefundOperations(t *testing.T) {
 				}
 			}(csDB)
 
-			refundValue := uint64(GetRandom(10000*4000, 10000*5000))
+			refundValue := GetRandom(10000*4000, 10000*5000)
 			csDB.AddRefund(refundValue)
 
 			if csDB.GetRefund() != refundValue {
@@ -550,7 +550,7 @@ func TestCarmenState_SetBalanceUsingBulkInsertion(t *testing.T) {
 
 			cbl.CreateAccount(addr)
 
-			newBalance := uint256.NewInt(uint64(GetRandom(1, 1000*5000)))
+			newBalance := uint256.NewInt(GetRandom(1, 1000*5000))
 			cbl.SetBalance(addr, newBalance)
 
 			err = cbl.Close()
@@ -600,7 +600,7 @@ func TestCarmenState_SetNonceUsingBulkInsertion(t *testing.T) {
 
 			cbl.CreateAccount(addr)
 
-			newNonce := uint64(GetRandom(1, 1000*5000))
+			newNonce := GetRandom(1, 1000*5000)
 
 			cbl.SetNonce(addr, newNonce)
 
@@ -773,7 +773,7 @@ func TestCarmenState_BulkloadOperations(t *testing.T) {
 				switch {
 				case operationType == 1:
 					// set balance
-					newBalance := uint256.NewInt(uint64(GetRandom(0, 1000*5000)))
+					newBalance := uint256.NewInt(GetRandom(0, 1000*5000))
 
 					cbl.SetBalance(account, newBalance)
 				case operationType == 2:
@@ -789,7 +789,7 @@ func TestCarmenState_BulkloadOperations(t *testing.T) {
 					cbl.SetState(account, key, value)
 				case operationType == 4:
 					// set nonce
-					newNonce := uint64(GetRandom(0, 1000*5000))
+					newNonce := GetRandom(0, 1000*5000)
 
 					cbl.SetNonce(account, newNonce)
 				default:
